Move abs_path to util
